Update UI components with role-based filtering and improvements
- Update list components to use role-based filtering
- Improve production order components
- Update analysis containers and goal sections
- Fix sidebar menu and login credentials components

+++ src/components/deleteButton/index.tsx
@@ -10,16 +10,35 @@ import { useLoading } from "@/hooks/useLoading";
 const DeleteButton = ({ uuid, refetch, endpoint }: { uuid: string; refetch?: () => void; endpoint: string }) => {
   const { setIsLoading } = useLoading();
 
+  async function handleUserDeactivation() {
+    setIsLoading(true);
+    try {
+      await api.delete(`users/deactivate/${uuid}`);
+
+      toast.success("Registro excluido com sucesso!");
+
+      if (refetch) refetch();
+    } catch (err) {
+      const error = err as Error;
+      toast.error(error.message);
+    }
+  }
+
   async function handleDelete(uuid: string) {
     setIsLoading(true);
+    console.log(`uuid: ${uuid}`);
+
+    if (endpoint === "users") {
+      await handleUserDeactivation();
+      return;
+    }
+
     try {
       await api.delete(`${endpoint}/${uuid}`);
 
       toast.success("Registro excluido com sucesso!");
 
-      if (refetch) {
-        refetch();
-      }
+      if (refetch) refetch();
     } catch (err) {
       toast.error((err as Error).message);
     } finally {
